SBI-24: Extract EvmChainConfig and EvmChainTokenConfig to separate files
Move nested records out of EvmChainIndexerFactory into their own
top-level files per project conventions. Rename TokenConfig to
EvmChainTokenConfig for clarity.
