+++ stablebridge-indexer/src/main/java/com/stablebridge/indexer/infrastructure/chain/evm/EvmChainConfig.java
@@ -1,0 +1,23 @@
+package com.stablebridge.indexer.infrastructure.chain.evm;
+
+import lombok.Builder;
+
+import java.time.Duration;
+import java.util.List;
+
+@Builder(toBuilder = true)
+public record EvmChainConfig(
+        String networkId,
+        String rpcUrl,
+        int rpcBatchSize,
+        boolean useBlockReceipts,
+        Duration rpcTimeout,
+        int maxRetries,
+        int rateLimitRps,
+        int rateLimitBurst,
+        boolean useFinalizedTag,
+        int minConfirmations,
+        boolean indexNativeTransfers,
+        int nativeDecimals,
+        List<EvmChainTokenConfig> tokenContracts
+) {}
