+++ stablebridge-indexer/src/main/java/com/stablebridge/indexer/infrastructure/chain/evm/EvmChainTokenConfig.java
@@ -1,0 +1,10 @@
+package com.stablebridge.indexer.infrastructure.chain.evm;
+
+import lombok.Builder;
+
+@Builder(toBuilder = true)
+public record EvmChainTokenConfig(
+        String address,
+        String symbol,
+        int decimals
+) {}
